音響: 波の枠を 16 → 32 本に / 満杯のとき敵の声は敵でない最古の波を追い出す
三校で、金属床を走る足音と追跡中の 3 体の声で 16 本が埋まり、敵の声が捨てられて
「敵が見えない」が起きうるため (2026-09-14)。

- AcousticField::kMaxWaves 16 -> 32。見通しマスク (FrontMask) を uint16 -> uint32、
  GPU 側を R16_UINT -> R32_UINT、MYE_ACOUSTIC_WAVE_SLOTS / RenderView::kAcousticWaveSlots を 32 に
- Emit の agentPriority: 満杯でも AgentSystem の声は、AgentBrain を持たない発音元のうち
  最古の波を追い出して立つ (敵の声どうしは追い出さない)
- SimSnapshot v21 (ACU 節の波スロット表が 32 本。本数の違う古い blob は拒む)
- selftest ALL PASS
- ★shot_verify の acoustic_forward / acoustic_deferred は golden 未更新で不一致
  (敵の位置の差。maxDiff 83 / 約 600 px)

+++ src/Engine/Engine/RenderSystem.cpp
@@ -121,7 +121,7 @@ void DumpAcousticVolume(GraphicsDevice& device, const AcousticVolumePass& pass,
     const float keepRaw = field.GlowKeepPerTick();
     const float keep = (keepRaw > 0.0f && keepRaw < 1.0f) ? keepRaw : acoustic::kGlowDecayPerTick;
     const AcousticField::FrontWave* fws = field.FrontWaves();
-    const std::vector<uint16_t>& mask = field.FrontMask();
+    const std::vector<uint32_t>& mask = field.FrontMask();
     for (uint32_t s = 0; s < AcousticField::kMaxWaves; ++s) {
         const AcousticField::FrontWave& fw = fws[s];
         if (fw.active == 0) {
@@ -1327,7 +1327,7 @@ bool RenderSystem::Render(World& world, GraphicsDevice& device, IRenderPath& pat
                 acousticSupplied_ = true;
                 // 2026-09-12「描画だけ円」: 見通しビットと波の表。マスクの転送に失敗したら
                 // SRV が null のまま = 従来 (残光だけ) の絵に静かに戻る
-                const std::vector<uint16_t>& front = acousticField->FrontMask();
+                const std::vector<uint32_t>& front = acousticField->FrontMask();
                 if (acousticFront && acousticField->FrontActive()
                     && static_cast<int64_t>(front.size()) == ag.CellCount()
                     && acousticPass_.UploadFront(device, front.data(), ag.dimX, ag.dimY, ag.dimZ,
